backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.11
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==3.2.14
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-spectacular==0.23.1
drf-spectacular-sidecar==2022.8.1
drf-yasg==1.21.3
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==1.7.0
importlib-resources==5.9.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.9.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-django==4.5.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
XlsxWriter==3.0.3
zipp==3.7.0
